requirements.txt
azure-eventhub
azure-identity